+++ ErsatzTV.Core/Metadata/FallbackMetadataProvider.cs
@@ -230,6 +230,7 @@ public class FallbackMetadataProvider : IFallbackMetadataProvider
             metadata.Title = match.Success
                 ? match.Groups[2].Value.Trim()
                 : Path.GetFileNameWithoutExtension(fileName);
+            metadata.Artists = new List<MusicVideoArtist>();
             metadata.Genres = new List<Genre>();
             metadata.Tags = new List<Tag>();
             metadata.Studios = new List<Studio>();